+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/messages/MessageListViewModel.kt
@@ -5,17 +5,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.getstream.sdk.chat.state.Copy
-import com.getstream.sdk.chat.state.Delete
-import com.getstream.sdk.chat.state.Flag
-import com.getstream.sdk.chat.state.MessageAction
-import com.getstream.sdk.chat.state.MessageMode
-import com.getstream.sdk.chat.state.MuteUser
-import com.getstream.sdk.chat.state.Normal
-import com.getstream.sdk.chat.state.React
-import com.getstream.sdk.chat.state.Reply
-import com.getstream.sdk.chat.state.Thread
-import com.getstream.sdk.chat.state.ThreadReply
 import com.getstream.sdk.chat.viewmodel.messages.getCreatedAtOrThrow
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.call.await
@@ -23,6 +12,15 @@ import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.User
+import io.getstream.chat.android.common.state.Copy
+import io.getstream.chat.android.common.state.Delete
+import io.getstream.chat.android.common.state.Flag
+import io.getstream.chat.android.common.state.MessageAction
+import io.getstream.chat.android.common.state.MessageMode
+import io.getstream.chat.android.common.state.MuteUser
+import io.getstream.chat.android.common.state.React
+import io.getstream.chat.android.common.state.Reply
+import io.getstream.chat.android.common.state.ThreadReply
 import io.getstream.chat.android.compose.handlers.ClipboardHandler
 import io.getstream.chat.android.compose.state.messages.MessagesState
 import io.getstream.chat.android.compose.state.messages.MyOwn
@@ -41,15 +39,24 @@ import io.getstream.chat.android.offline.model.ConnectionState
 import io.getstream.chat.android.offline.thread.ThreadController
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import java.util.Date
 import java.util.concurrent.TimeUnit
 
 /**
  * ViewModel responsible for handling all the business logic & state for the list of messages.
+ *
+ * @param chatClient Used to connect to the API.
+ * @param chatDomain Used to connect to the API and fetch the domain status.
+ * @param channelId Used to connect to the required channel.
+ * @param messageLimit The limit of messages being fetched with each page od data.
+ * @param enforceUniqueReactions Enables or disables unique message reactions per user.
+ * @param clipboardHandler Used to copy data from message actions to the clipboard.
  */
 public class MessageListViewModel(
     public val chatClient: ChatClient,
@@ -69,19 +76,19 @@ public class MessageListViewModel(
         get() = if (isInThread) threadMessagesState else messagesState
 
     /**
-     * State of the screen, for the [Normal] [messageMode].
+     * State of the screen, for [MessageMode.Normal].
      */
     private var messagesState: MessagesState by mutableStateOf(MessagesState())
 
     /**
-     * State of the screen, for the [Thread] [messageMode].
+     * State of the screen, for [MessageMode.MessageThread].
      */
     private var threadMessagesState: MessagesState by mutableStateOf(MessagesState())
 
     /**
-     * Holds the current [MessageMode] that's used for the messages list. [Normal] by default.
+     * Holds the current [MessageMode] that's used for the messages list. [MessageMode.Normal] by default.
      */
-    public var messageMode: MessageMode by mutableStateOf(Normal)
+    public var messageMode: MessageMode by mutableStateOf(MessageMode.Normal)
         private set
 
     /**
@@ -117,9 +124,8 @@ public class MessageListViewModel(
     /**
      * Gives us information about the online state of the device.
      */
-    @Deprecated("Use connectionState instead")
-    public val isOnline: StateFlow<Boolean>
-        get() = chatDomain.online
+    public val isOnline: Flow<Boolean>
+        get() = chatDomain.connectionState.map { it == ConnectionState.CONNECTED }
 
     /**
      * Gives us information about the logged in user state.
@@ -375,7 +381,7 @@ public class MessageListViewModel(
     public fun loadMore() {
         val messageMode = messageMode
 
-        if (messageMode is Thread) {
+        if (messageMode is MessageMode.MessageThread) {
             threadMessagesState = threadMessagesState.copy(isLoadingMore = true)
             chatDomain.threadLoadMore(channelId, messageMode.parentMessage.id, messageLimit)
                 .enqueue()
@@ -406,7 +412,7 @@ public class MessageListViewModel(
      * @param message The selected message with a thread.
      */
     public fun openMessageThread(message: Message) {
-        this.messageMode = Thread(message)
+        this.messageMode = MessageMode.MessageThread(message)
 
         loadThread(message)
     }
@@ -465,7 +471,7 @@ public class MessageListViewModel(
      * @param parentMessage The message with the thread we want to observe.
      */
     private fun loadThread(parentMessage: Message) {
-        messageMode = Thread(parentMessage)
+        messageMode = MessageMode.MessageThread(parentMessage)
 
         chatDomain.getThread(channelId, parentMessage.id).enqueue { result ->
             if (result.isSuccess) {
@@ -473,7 +479,7 @@ public class MessageListViewModel(
 
                 observeThreadMessages(controller)
             } else {
-                messageMode = Normal
+                messageMode = MessageMode.Normal
             }
         }
     }
@@ -512,7 +518,7 @@ public class MessageListViewModel(
      * @return A list of [MessageItem]s, each containing a position.
      */
     private fun groupMessages(messages: List<Message>): List<MessageListItem> {
-        val parentMessageId = (messageMode as? Thread)?.parentMessage?.id
+        val parentMessageId = (messageMode as? MessageMode.MessageThread)?.parentMessage?.id
         val currentUser = user.value
         val groupedMessages = mutableListOf<MessageListItem>()
 
@@ -612,7 +618,7 @@ public class MessageListViewModel(
      * It also cancels the [threadJob] to clean up resources.
      */
     public fun leaveThread() {
-        messageMode = Normal
+        messageMode = MessageMode.Normal
         messagesState = messagesState.copy(selectedMessage = null)
         threadMessagesState = MessagesState()
         lastSeenThreadMessage = null